pyproject.toml
[project]
name = "eye-of-horus"
version = "0.1.0"
description = ""
authors = [
    {name = "gnana-sai",email = "Gnana.sai@shipsy.io"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-genai (>=1.11.0,<2.0.0)",
    "tenacity (>=9.1.2,<10.0.0)",
    "google-cloud (>=0.34.0,<0.35.0)",
    "fastapi (>=0.115.12,<0.116.0)",
    "google-api-core (>=2.24.2,<3.0.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "boto3 (>=1.38.2,<2.0.0)",
    "uvicorn (>=0.34.2,<0.35.0)",
    "jinja2 (>=3.1.6,<4.0.0)",
    "google-auth (>=2.39.0,<3.0.0)",
    "google-auth-oauthlib (>=1.2.2,<2.0.0)",
    "google-auth-httplib2 (>=0.2.0,<0.3.0)",
    "google-api-python-client (>=2.168.0,<3.0.0)",
    "pydub (>=0.25.1,<0.26.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
